Responsiveness of showblog and search utility in blog manager

+++ public/Blog/Blogmanager.js
@@ -1,12 +1,15 @@
 var url = 'https://codingclubnitm.herokuapp.com/api/v1/blog/'
-var deleteblog = document.querySelector('#delete');
 
 function deleteblog(id) {
     console.log('Fired')
-    fetch(url + id, {
-        method: 'DELETE',
-    }).then(res => {
-        window.location.reload()
-        return res.json()
-    });
+    var result = confirm("Want to delete?");
+    if (result) {
+        fetch(url + id, {
+            method: 'DELETE',
+        }).then(res => {
+            window.location.reload()
+            return res.json()
+        });
+    }
+
 }
